+++ src/ride.h
@@ -1,7 +1,7 @@
 #pragma once
 
 typedef struct {
-  float distance;
+  int distance;
   char direction;
   char cue[128];
 } ride_entry_t;
